requirements.txt
boto3 >= 1.21.4
colorama
pyyaml